pyproject.toml
[tool.poetry]
name = "rework-document-update"
version = "0.1.0"
description = "Reworked automated document updates based on software changes"
authors = ["Your Name <your.email@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
grpcio = "^1.59.0"
grpcio-tools = "^1.59.0"
python-arango = "^7.7.0"
python-docx = "^1.1.0"
# openai = "^1.3.0"  # Removed - now using Ollama for local LLM
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
structlog = "^23.2.0"
protobuf = "^5.26.0"
requests = "^2.31.0"
GitPython = "^3.1.40"
flask = "^3.0.0"
# Additional dependencies for the documentation system
openpyxl = "^3.1.0"  # For Excel file handling
python-magic = "^0.4.27"  # For file type detection

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.11.0"
isort = "^5.12.0"

[tool.poetry.scripts]
rework-server = "main:serve"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

